Move `get_parameter_values` and `fetch_parameters` to shared ena_std_lib
